Bug-fix: dict_compiler::use_preset_vocabulary option not initialized to false.
git-svn-id: https://rimeime.googlecode.com/svn/trunk/librime@461 b90aed45-4297-9963-ff9d-98b604ac09a9

+++ src/dict/prism.cc
@@ -35,7 +35,7 @@ namespace rime {
 
 SpellingAccessor::SpellingAccessor(prism::SpellingMap* spelling_map, int spelling_id)
     : spelling_id_(spelling_id), iter_(NULL), end_(NULL) {
-  if (spelling_map && spelling_id < spelling_map->size) {
+  if (spelling_map && spelling_id < static_cast<int>(spelling_map->size)) {
     iter_ = spelling_map->at[spelling_id].begin();
     end_ = spelling_map->at[spelling_id].end();
   }
